// test-user-end/Helpers/Types.tsx
export type IRound = {
  dec: number;
  number: number;
};